refactor: Move getWorkbookWithAuthor to service layer (#3193)
