+++ src/routes/workbooks/[slug]/+page.server.ts
@@ -4,10 +4,10 @@ import { Roles } from '$lib/types/user';
 import type { TaskResult } from '$lib/types/task';
 
 import * as taskResultsCrud from '$lib/services/task_results';
+import { getWorkbookWithAuthor } from '$features/workbooks/services/workbooks';
 import * as action from '$lib/actions/update_task_result';
 
 import { getLoggedInUser, isAdmin, canRead } from '$lib/utils/authorship';
-import { getWorkbookWithAuthor } from '$features/workbooks/utils/workbook';
 import { FORBIDDEN } from '$lib/constants/http-response-status-codes';
 
 export async function load({ locals, params }) {
